+++ include/sick_safetyscanners_base/AsioCompat.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <boost/asio.hpp>
-#include <boost/asio/deadline_timer.hpp>
 #include <boost/version.hpp>
 
 namespace boost {