+++ include/sick_safetyscanners_base/communication/UDPClient.h
@@ -40,6 +40,7 @@
 #include <iostream>
 
 #include "sick_safetyscanners_base/AsioCompat.hpp"
+#include <boost/asio/steady_timer.hpp>
 
 #include "sick_safetyscanners_base/Types.h"
 #include "sick_safetyscanners_base/datastructure/PacketBuffer.h"
@@ -144,7 +145,7 @@ private:
   boost::asio::ip::udp::socket m_socket;
   types::PacketHandler m_packet_handler;
   datastructure::PacketBuffer::ArrayBuffer m_recv_buffer;
-  boost::asio::deadline_timer m_deadline;
+  boost::asio::steady_timer m_deadline;
 
   /*!
    * \brief A function to check internal deadline constraints on connect, receive and send