+++ src/communication/TCPClient.cpp
@@ -46,8 +46,8 @@
 namespace sick {
 namespace communication {
 
-using boost::asio::deadline_timer;
 using boost::asio::ip::tcp;
+using boost::asio::steady_timer;
 using boost::lambda::_1;
 using boost::lambda::_2;
 using boost::lambda::bind;
@@ -63,7 +63,7 @@ TCPClient::TCPClient(sick::types::ip_address_t server_ip, sick::types::port_t se
 {
   // Since synchronous timeout on socket.connect() are not available with boost::asio, this seems to
   // be the most elegant way eventhough asynchronous.
-  m_deadline.expires_at(boost::posix_time::pos_infin);
+  m_deadline.expires_at(steady_timer::time_point::max());
   checkDeadline();
 }
 
@@ -72,7 +72,7 @@ void TCPClient::checkDeadline()
   // Check whether the deadline has passed. We compare the deadline against
   // the current time since a new asynchronous operation may have moved the
   // deadline before this actor had a chance to run.
-  if (m_deadline.expires_at() <= deadline_timer::traits_type::now())
+  if (m_deadline.expiry() <= steady_timer::clock_type::now())
   {
     // The deadline has passed. The socket is closed so that any outstanding
     // asynchronous operations are cancelled. This allows the blocked
@@ -82,7 +82,7 @@ void TCPClient::checkDeadline()
 
     // There is no longer an active deadline. The expiry is set to positive
     // infinity so that the actor takes no action until a new deadline is set.
-    m_deadline.expires_at(boost::posix_time::pos_infin);
+    m_deadline.expires_at(steady_timer::time_point::max());
   }
 
   // Put the actor back to sleep.
@@ -94,7 +94,14 @@ void TCPClient::connect(sick::types::time_duration_t timeout)
   auto remote_endpoint         = boost::asio::ip::tcp::endpoint(m_server_ip, m_server_port);
   boost::system::error_code ec = boost::asio::error::would_block;
 
-  m_deadline.expires_from_now(timeout);
+  if (timeout == sick::types::infinite_timeout)
+  {
+    m_deadline.expires_at(steady_timer::time_point::max());
+  }
+  else
+  {
+    m_deadline.expires_after(timeout);
+  }
   m_socket.async_connect(remote_endpoint, var(ec) = _1);
 
   do
@@ -148,7 +155,14 @@ sick::datastructure::PacketBuffer TCPClient::receive(sick::types::time_duration_
 {
   boost::system::error_code ec = boost::asio::error::would_block;
 
-  m_deadline.expires_from_now(timeout);
+  if (timeout == sick::types::infinite_timeout)
+  {
+    m_deadline.expires_at(steady_timer::time_point::max());
+  }
+  else
+  {
+    m_deadline.expires_after(timeout);
+  }
 
   std::size_t bytes_recv = 0;