+++ src/communication/UDPClient.cpp
@@ -53,8 +53,8 @@
 namespace sick {
 namespace communication {
 
-using boost::asio::deadline_timer;
 using boost::asio::ip::tcp;
+using boost::asio::steady_timer;
 using boost::lambda::_1;
 using boost::lambda::_2;
 using boost::lambda::bind;
@@ -67,7 +67,7 @@ UDPClient::UDPClient(boost::asio::io_service& io_service, sick::types::port_t se
   , m_recv_buffer()
   , m_deadline(io_service)
 {
-  m_deadline.expires_at(boost::posix_time::pos_infin);
+  m_deadline.expires_at(steady_timer::time_point::max());
   checkDeadline();
 }
 
@@ -88,7 +88,7 @@ UDPClient::UDPClient(boost::asio::io_service& io_service,
   }
   m_socket.set_option(boost::asio::ip::multicast::join_group(host_ip, interface_ip));
 
-  m_deadline.expires_at(boost::posix_time::pos_infin);
+  m_deadline.expires_at(steady_timer::time_point::max());
   checkDeadline();
 }
 
@@ -99,7 +99,7 @@ void UDPClient::checkDeadline()
   // Check whether the deadline has passed. We compare the deadline against
   // the current time since a new asynchronous operation may have moved the
   // deadline before this actor had a chance to run.
-  if (m_deadline.expires_at() <= deadline_timer::traits_type::now())
+  if (m_deadline.expiry() <= steady_timer::clock_type::now())
   {
     // The deadline has passed. The socket is closed so that any outstanding
     // asynchronous operations are cancelled. This allows the blocked
@@ -110,7 +110,7 @@ void UDPClient::checkDeadline()
     // There is no longer an active deadline. The expiry is set to positive
     // infinity so that the actor takes no action until a new deadline is set.
     m_deadline.cancel();
-    m_deadline.expires_at(boost::posix_time::pos_infin);
+    m_deadline.expires_at(steady_timer::time_point::max());
   }
 
   // Put the actor back to sleep.
@@ -148,7 +148,14 @@ void UDPClient::stop()
 sick::datastructure::PacketBuffer UDPClient::receive(sick::types::time_duration_t timeout)
 {
   boost::system::error_code ec = boost::asio::error::would_block;
-  m_deadline.expires_from_now(timeout);
+  if (timeout == sick::types::infinite_timeout)
+  {
+    m_deadline.expires_at(steady_timer::time_point::max());
+  }
+  else
+  {
+    m_deadline.expires_after(timeout);
+  }
 
   std::size_t bytes_recv = 0;
   m_socket.async_receive_from(